requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pymysql==1.1.1
cryptography==44.0.0
python-dotenv==1.0.1
jinja2==3.1.5
itsdangerous==2.2.0
python-multipart==0.0.20
apscheduler==3.10.4
httpx==0.28.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
